requirements.txt
aniso8601==10.0.1
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
libpass==1.9.1.post0
MarkupSafe==3.0.2
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
WTForms==3.2.1
